+++ pid-be/app/models/entities/Appointment.py
@@ -57,6 +57,7 @@ class Appointment:
     @staticmethod
     def delete_by_id(id):
         db.collection("appointments").document(id).delete()
+        
 
     def create(self):
         id = db.collection("appointments").document().id
