--- pid-ui/src/app/dashboard-redirect/dashboard-redirect.module.css
@@ -1,294 +1,0 @@
-/* Estilos para la página de Dashboard */
-.admin {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    height: 100vh;
-    text-align: center;
-    background-color: #fff;
-    padding: 0;
-    margin-top: 100px;
-}
-
-/* Estilos para el encabezado */
-.header {
-    background-color: rgb(238, 238, 238); /* Color de fondo del encabezado */
-    padding: 1rem; /* Espaciado del encabezado */
-    text-align: center;
-    position: fixed; /* Establece la posición como fija */
-    top: 0; /* Pega el encabezado al borde superior */
-    left: 0; /* Alinea el encabezado a la izquierda */
-    right: 0; /* Alinea el encabezado a la derecha */
-    z-index: 999; /* Ajusta el valor de z-index para superponer sobre el contenido */
-    display: flex; /* Activa el diseño flexible para centrar el contenido verticalmente */
-    justify-content: center; /* Centra el contenido horizontalmente */
-    align-items: center; /* Centra el contenido verticalmente */
-    height: 100px; /* Fija una altura fija para el encabezado */
-}
-
-/* Estilos para los mensajes de error */
-.error-message {
-    color: red;
-    font-size: 1rem;
-    margin-bottom: 1rem;
-}
-
-/* Estilos para el logo */
-.logo {
-    width: 250px; /* Tamaño del logo */
-    height: auto;
-    cursor: pointer;
-}
-
-.logout-icon {
-    width: 35px;
-    height: auto;
-    cursor: pointer;
-    position: absolute;
-    right: 2rem;
-}
-.logout-icon:hover {
-    filter: brightness(0.8);
-}
-
-/* Estilos para el título de la página de inicio de sesión */
-.title {
-    font-size: 2rem;
-    margin-bottom: 1rem;
-    color: #333;
-}
-
-/* Estilos para el subtítulo "Iniciar Sesión" */
-.subtitle {
-    font-size: 1.5rem; /* Tamaño del subtítulo */
-    color: #333;
-}
-
-.subtitle:disabled {
-    color: #5d5d5d;
-}
-
-/* Estilos para la barra superior de navegación por pestañas */
-.tab-bar {
-    display: flex;
-    background-color: rgb(238, 238, 238);
-    text-align: center;
-    position: fixed;
-    top: 100px; /* Alinea la barra debajo del encabezado */
-    left: 0;
-    right: 0;
-    z-index: 999;
-    justify-content: center;
-    align-items: center;
-    height: 50px; /* Altura fija para la barra */
-    width: 100%; /* Ancho del 100% para ocupar toda la pantalla */
-    padding: 2rem;
-    box-shadow: inset 0 -10px 10px -10px rgba(0, 0, 0, 0.1);
-}
-
-/* Estilos para las pestañas */
-.tab {
-    flex: 1;
-    text-align: center;
-    cursor: pointer;
-    padding: 10px 20px; /* Añade espaciado a las pestañas */
-    color: #333;
-    border-radius: 6px;
-    background-color: #cacaca;
-    margin: 10px 10px 10px 10px;
-}
-
-.tab_disabled {
-    flex: 1;
-    text-align: center;
-    cursor: not-allowed;
-    padding: 10px 20px; /* Añade espaciado a las pestañas */
-    color: #5d5d5d;
-    border-radius: 6px;
-    background-color: #d7d7d7;
-    margin: 10px 10px 10px 10px;
-}
-
-/* Estilos para el contenido de la pestaña activa */
-.tab-content {
-    flex: 1;
-    padding: 2rem;
-    width: 100%;
-}
-
-/* Estilos para el formulario de selección de especialidad y doctor */
-.form {
-    flex: 1;
-    align-self: flex-start;
-    display: flex;
-    width: 60rem;
-    max-width: 100%;
-    margin: 0 auto;
-    background-color: #fff;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-    text-align: left;
-    width: 100%; /* Ancho del 100% para ocupar toda la pantalla */
-    padding: 2rem;
-}
-
-/* Estilos para los grupos de formulario */
-.form {
-    text-align: left;
-    margin-bottom: 1rem;
-    display: flex;
-    flex-direction: column; /* Apila los elementos verticalmente */
-}
-
-/* Estilos para las etiquetas de formulario */
-.form label {
-    font-size: 1rem;
-    margin-bottom: 0.5rem;
-    color: #333;
-}
-
-/* Estilos para los campos de selección (select) */
-.form select {
-    width: 100%;
-    padding: 0.5rem;
-    font-size: 1rem;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    color: #000;
-    background-color: #fff;
-    appearance: none;
-    outline: none;
-    cursor: pointer;
-    margin-bottom: 2rem;
-}
-
-.form option {
-    color: #000;
-}
-
-.form input {
-    width: 100%;
-    padding: 0.5rem;
-    font-size: 1rem;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    color: #000;
-    background-color: #fff;
-    appearance: none;
-    outline: none;
-    cursor: pointer;
-    margin-bottom: 2rem;
-}
-
-/* Estilos para el botón de envío */
-.submit-button {
-    padding: 0.5rem 1rem;
-    background-color: #007bff;
-    color: #fff;
-    text-decoration: none;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    transition: background-color 0.3s;
-    width: 100%;
-    text-align: center;
-    margin-top: 1rem;
-}
-
-.submit-button:hover {
-    background-color: #0056b3;
-}
-
-/* Estilos para el botón deshabilitado */
-.disabled-button {
-    background-color: #ccc;
-    cursor: not-allowed;
-}
-
-/* Estilos para el pie de página de la página de inicio */
-.page-footer {
-    font-size: 0.8rem;
-    margin-top: 2rem;
-    margin-bottom: 2rem;
-    color: #777;
-    position: relative;
-    bottom: 0;
-}
-/* Estilos para la sección de "Mis Proximos Turnos" */
-.appointments-section {
-    margin-top: 2rem;
-    background-color: #f5f5f5; /* Fondo gris claro */
-    padding: 1rem;
-    border-radius: 5px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1); /* Sombra suave */
-}
-
-.appointment-buttons-container {
-    display: flex; /* Activa el diseño flexible */
-    justify-content: flex-end; /* Alinea los elementos al final (derecha) del contenedor */
-}
-
-.appointment {
-    margin-bottom: 1rem;
-    padding: 1rem;
-    background-color: #fff; /* Fondo blanco */
-    border-radius: 5px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.appointment p {
-    margin: 0.1rem;
-    font-size: 1.2rem;
-    color: #333;
-}
-
-.stantard-button {
-    background-color: #007bff;
-    padding: 0.5rem 1rem;
-    color: #fff;
-    text-decoration: none;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    transition: background-color 0.3s;
-    margin-right: 10px;
-}
-
-.stantard-button:hover {
-    background-color: #0056b3;
-}
-
-/* Estilos para el botón de eliminación */
-.deny-button {
-    background-color: #ff3333; /* Cambia el color para el botón de eliminar */
-    padding: 0.5rem 1rem;
-    color: #fff;
-    text-decoration: none;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    transition: background-color 0.3s;
-    margin-right: 10px; /* Espaciado a la derecha para separar los botones */
-}
-.deny-button:hover {
-    background-color: #cc0000; /* Cambia el color al pasar el mouse */
-}
-
-/* Estilos para el botón de edición*/
-.approve-button {
-    background-color: #00a010; /* Cambia el color para el botón de editar */
-    padding: 0.5rem 1rem;
-    color: #fff;
-    text-decoration: none;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    transition: background-color 0.3s;
-    margin-right: 10px; /* Espaciado a la derecha para separar los botones */
-}
-
-.approve-button:hover {
-    background-color: #008a0a; /* Cambia el color al pasar el mouse */
-}
