+++ pid-ui/src/app/landing/page.js
@@ -5,6 +5,8 @@ import Link from "next/link";
 import styles from "./landing.module.css";
 import { useRouter } from "next/navigation";
 import axios from "axios";
+import { Footer, HeaderSlim } from "../components/header";
+import userCheck from "../components/userCheck";
 
 const Landing = () => {
     const [email, setEmail] = useState("");
@@ -33,12 +35,15 @@ const Landing = () => {
             );
             localStorage.setItem("token", response.data.token);
             sessionStorage.setItem("user", JSON.stringify(response.data.user));
-            router.push("/dashboard-redirect");
+            userCheck(router);
+
+            //router.push("/dashboard-redirect");
         } catch (error) {
             setError("Error al iniciar sesión: " + error.response.data.detail);
 
             if (error.response.data.detail == "User has already logged in") {
-                router.push("/dashboard-redirect");
+                userCheck(router);
+                // router.push("/dashboard-redirect");
             }
 
             // Verificar si el elemento .error-message está presente en el DOM
@@ -52,15 +57,7 @@ const Landing = () => {
 
     return (
         <div className={styles["login-page"]}>
-            <header className={styles["header"]}>
-                <Link href="/">
-                    <img
-                        src="/logo.png"
-                        alt="logo"
-                        className={styles["logo"]}
-                    />
-                </Link>
-            </header>
+            <HeaderSlim />
             <form className={styles["form"]} onSubmit={handleLogin}>
                 <div className={styles["title"]}>¡Bienvenido!</div>
                 <div className={styles["subtitle"]}>Iniciar Sesion</div>
@@ -97,10 +94,7 @@ const Landing = () => {
                     </Link>
                 </div>
             </form>
-
-            <footer>
-                <p>Derechos de autor © 2023 KMK</p>
-            </footer>
+            <Footer />
         </div>
     );
 };
